templates: extract SessionTemplateGenerator from WPF assembly
Quantum-graph behavior moves to a QuantumGraphTemplateGenerator decorator
in Stride.Core.Assets.Quantum, applied at registration via
QuantumTemplateRegistration.Register. CLI/headless flows use the bare base.
